Helper methods tests moved from SeasonStats to HelperMethodsTest.
